`getAllRepoStats` for wiget
