+++ pages/widget.jsx
@@ -1,7 +1,8 @@
 import { useEffect } from "react";
 import Home from "./";
+import getAllRepoStats from "../github-api/get-all-repo-stats";
 
-export default function Widget() {
+export default function Widget({ projectsData }) {
   useEffect(() => {
     const interval = setInterval(() => {
       const height = document.documentElement.scrollHeight;
@@ -12,7 +13,22 @@ export default function Widget() {
       clearInterval(interval);
     };
   }, []);
-  return <Home isWidgetVersion />;
+
+  return (
+    <Home
+      projectsData={projectsData}
+      isWidgetVersion
+    />
+  );
+}
+
+export async function getStaticProps() {
+  return {
+    props: {
+      projectsData: await getAllRepoStats()
+    },
+    revalidate: 3600
+  }
 }
 
 Widget.theme = "dark";
